requirements.txt
python-telegram-bot[rate-limiter]>=21.0,<22
python-dotenv>=1.0.0
